Split tests/fixtures into separate files
