+++ test/test_semanticnet.py
@@ -1,79 +1,9 @@
+import os
 import pytest
 import semanticnet as sn
-import uuid
-import os
 import time
-import json
+import uuid
 
-### FIXTURES ###
-@pytest.fixture
-def uuid_str():
-    return '3caaa8c09148493dbdf02c574b95526c'
-
-@pytest.fixture
-def uuid_obj():
-    return uuid.UUID('3caaa8c09148493dbdf02c574b95526c')
-
-@pytest.fixture
-def graph():
-    return sn.Graph()
-
-@pytest.fixture
-def populated_graph():
-    g = sn.Graph()
-    a = g.add_node({"type": "A"}, '3caaa8c09148493dbdf02c574b95526c')
-    b = g.add_node({"type": "B"}, '2cdfebf3bf9547f19f0412ccdfbe03b7')
-    c = g.add_node({"type": "C"}, '3cd197c2cf5e42dc9ccd0c2adcaf4bc2')
-    g.add_edge(a, b, {"type": "normal"}, '5f5f44ec7c0144e29c5b7d513f92d9ab')
-    g.add_edge(a, c, {"type": "normal"}, '7eb91be54d3746b89a61a282bcc207bb')
-    g.add_edge(b, c, {"type": "irregular"}, 'c172a3599b7d4ef3bbb688277276b763')
-    return g
-
-@pytest.fixture
-def populated_digraph():
-    g = sn.DiGraph()
-    a = g.add_node({"type": "A"}, '3caaa8c09148493dbdf02c574b95526c')
-    b = g.add_node({"type": "B"}, '2cdfebf3bf9547f19f0412ccdfbe03b7')
-    c = g.add_node({"type": "C"}, '3cd197c2cf5e42dc9ccd0c2adcaf4bc2')
-    g.add_edge(a, b, {"type": "normal"}, '5f5f44ec7c0144e29c5b7d513f92d9ab')
-    g.add_edge(b, a, {"type": "normal"}, 'f3674fcc691848ebbd478b1bfb3e84c3')
-    g.add_edge(a, c, {"type": "normal"}, '7eb91be54d3746b89a61a282bcc207bb')
-    g.add_edge(b, c, {"type": "irregular"}, 'c172a3599b7d4ef3bbb688277276b763')
-    return g
-
-@pytest.fixture
-def test_output():
-    graph = sn.Graph()
-
-    a = graph.add_node({"label" : "A"}, '6cf546f71efe47578f7a1400871ef6b8')
-    b = graph.add_node({"label" : "B"}, 'bcb388bb24a74d978fa2006ed278b2fe')
-    c = graph.add_node({"label" : "C"}, 'd6523f4f9d5240d2a92e341f4ca00a78')
-
-    graph.add_edge(a, b, {"type" : "belongs"}, 'ff8a8a8093cf436aa3b0127c71ddc11d')
-    graph.add_edge(b, c, {"type" : "owns"}, '081369f6197b467abe97b3efe8cc4640')
-    graph.add_edge(c, a, {"type" : "has"}, 'b3a245098d5d482f893c6d63606c7e91')
-
-    graph.save_json("test_output.json")
-
-    with open("test_output.json") as f:
-        jsonObj = json.load(f)
-
-    return jsonObj
-
-@pytest.fixture
-def correct_output():
-    with open(os.path.join(os.path.dirname(os.path.realpath(__file__)), "test_output_correct.json")) as f:
-        jsonObj = json.load(f)
-
-    return jsonObj
-
-@pytest.fixture
-def correct_output_graph():
-    g = sn.Graph()
-    g.load_json(os.path.join(os.path.dirname(os.path.realpath(__file__)), "test_output_correct.json"))
-    return g
-
-### TESTS ###
 def test__create_uuid(graph):
     id_ = graph._create_uuid()
     assert id_.__class__.__name__ == 'UUID'
@@ -99,75 +29,6 @@ def test_add_node(graph):
     node = graph.get_node_attributes(a)
     assert "type" in node
     assert node["type"] == "A"
-
-def test_add_node_with_cache(populated_graph):
-    populated_graph.cache_nodes_by("type")
-
-    ### get all nodes with attr "type"
-    input_ = populated_graph.get_nodes_by_attr("type")
-    output = {
-        "A": [
-            {
-                "id": uuid.UUID('3caaa8c0-9148-493d-bdf0-2c574b95526c'),
-                "type": "A"
-            }
-        ],
-        "B": [{
-            "id": uuid.UUID('2cdfebf3-bf95-47f1-9f04-12ccdfbe03b7'),
-            'type': 'B'
-        }],
-        "C": [{
-            "id": uuid.UUID('3cd197c2-cf5e-42dc-9ccd-0c2adcaf4bc2'),
-            'type': 'C'
-        }]
-    }
-
-    assert input_ == output
-
-    # Add a new node of the same type as another. Should be in the cache.
-    populated_graph.add_node({"type": "A"}, '2b673235a0b94935ab8b6b9de178d341')
-    input_ = populated_graph.get_nodes_by_attr("type", "A")
-    output = [
-        {
-            "id": uuid.UUID('3caaa8c0-9148-493d-bdf0-2c574b95526c'),
-            "type": "A"
-        },
-        {
-            "id": uuid.UUID('2b673235a0b94935ab8b6b9de178d341'),
-            "type": "A"
-        }
-    ]
-
-    assert input_ == output
-
-    # add a new node with an attribute that we are not caching.
-    # Should NOT be in the cache
-    populated_graph.add_node({"label": "test"}, 'fa02d5e82ed54baf828558c70317f20e')
-    input_ = populated_graph._node_cache
-    output = {
-        "type": {
-            "A": [
-                {
-                    "id": uuid.UUID('3caaa8c0-9148-493d-bdf0-2c574b95526c'),
-                    "type": "A"
-                },
-                {
-                    "id": uuid.UUID('2b673235a0b94935ab8b6b9de178d341'),
-                    "type": "A"
-                }
-            ],
-            "B": [{
-                "id": uuid.UUID('2cdfebf3-bf95-47f1-9f04-12ccdfbe03b7'),
-                'type': 'B'
-            }],
-            "C": [{
-                "id": uuid.UUID('3cd197c2-cf5e-42dc-9ccd-0c2adcaf4bc2'),
-                'type': 'C'
-            }]
-        }
-    }
-
-    assert input_ == output
 
 def test_get_node(populated_graph):
     assert (
@@ -229,18 +90,6 @@ def test_set_node_attribute(populated_graph):
     with pytest.raises(sn.GraphException):
         populated_graph.set_node_attribute('3caaa8c09148493dbdf02c574b95526c', 'id',
             '3caaa8c09148493dbdf02c57deadbeef')
-
-def test_set_node_attribute_with_cache(populated_graph):
-    populated_graph.cache_nodes_by("type")
-    populated_graph.set_node_attribute('3caaa8c09148493dbdf02c574b95526c', 'type', 'B')
-
-    a_nodes = populated_graph.get_nodes_by_attr("type", "A")
-    b_nodes = populated_graph.get_nodes_by_attr("type", "B")
-
-    node_a = populated_graph.get_node('3caaa8c09148493dbdf02c574b95526c')
-
-    assert node_a not in a_nodes
-    assert node_a in b_nodes
 
 def test_add_edge(graph):
     a = graph.add_node({"type": "A"})
@@ -433,19 +282,6 @@ def test_remove_digraph_node(populated_digraph):
     assert edge_b_a_id not in edges
     assert edge_a_c_id not in edges
     assert edge_b_c_id in edges
-
-def test_remove_node_with_cache(populated_graph):
-    populated_graph.add_node({"type": "A"}, '2b673235a0b94935ab8b6b9de178d341')
-    populated_graph.cache_nodes_by("type")
-    populated_graph.remove_node('3caaa8c09148493dbdf02c574b95526c')
-
-    input_ = populated_graph.get_nodes_by_attr("type", "A")
-    output = [{
-        "id": uuid.UUID('2b673235a0b94935ab8b6b9de178d341'),
-        "type": "A"
-    }]
-
-    assert input_ == output
 
 def test_save_json(test_output, correct_output):
     assert test_output["timeline"] == correct_output["timeline"]
